pyproject.toml
[project]
name = "deepresearch-mcp"
version = "0.1.0"
description = "MCP server plus Agents SDK runner for OpenAI Deep Research"
requires-python = ">=3.10"
dependencies = [
  "fastmcp>=0.4.0",
  "openai>=1.59.0",
  "openai-agents>=0.2.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "uvicorn>=0.32.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

